+++ backend/cmd/update.go
@@ -1,8 +1,10 @@
 package cmd
 
 import (
+	"fmt"
+
+	"github.com/ananthakumaran/paisa/pkg/database"
 	"github.com/ananthakumaran/paisa/pkg/model"
-	"github.com/ananthakumaran/paisa/pkg/utils"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -16,10 +18,10 @@ var (
 var updateCmd = &cobra.Command{
 	Use:   "update",
 	Short: "Sync journal data",
-	Run: func(cmd *cobra.Command, args []string) {
-		db, err := utils.OpenDB()
+	RunE: func(cmd *cobra.Command, args []string) error {
+		db, err := database.Initialize()
 		if err != nil {
-			log.Fatal(err)
+			return err
 		}
 
 		syncAll := !updateJournal && !updateCommodities && !updatePortfolios
@@ -27,7 +29,7 @@ var updateCmd = &cobra.Command{
 		if syncAll || updateJournal {
 			message, err := model.SyncJournal(db)
 			if err != nil {
-				log.Fatal(message)
+				return fmt.Errorf("sync journal: %s", message)
 			}
 		}
 
@@ -48,6 +50,7 @@ var updateCmd = &cobra.Command{
 				log.Warn("Failed to sync CII: ", err)
 			}
 		}
+		return nil
 	},
 }