+++ backend/pkg/model/db_benchmark_test.go
@@ -1,29 +1,26 @@
 package model
 
 import (
+	"context"
 	"path/filepath"
 	"testing"
 	"time"
 
 	"github.com/ananthakumaran/paisa/pkg/config"
+	"github.com/ananthakumaran/paisa/pkg/database"
 	"github.com/ananthakumaran/paisa/pkg/model/posting"
 	"github.com/ananthakumaran/paisa/pkg/model/price"
 	"github.com/ananthakumaran/paisa/pkg/query"
 	"github.com/stretchr/testify/require"
-	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
 )
 
 func setupBenchDB(b *testing.B, numPostings int) *gorm.DB {
 	b.Helper()
 	dir := b.TempDir()
 	dbPath := filepath.Join(dir, "bench.db")
-	db, err := gorm.Open(sqlite.Open(dbPath), &gorm.Config{
-		Logger: logger.Discard,
-	})
+	db, err := database.InitializePath(context.Background(), dbPath)
 	require.NoError(b, err)
-	require.NoError(b, AutoMigrate(db))
 
 	if numPostings > 0 {
 		postings := GenerateSyntheticPostings(numPostings)
@@ -38,8 +35,7 @@ func BenchmarkPosting_UpsertAll_1k(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		b.StopTimer()
 		dir := b.TempDir()
-		db, _ := gorm.Open(sqlite.Open(filepath.Join(dir, "bench.db")), &gorm.Config{Logger: logger.Discard})
-		_ = AutoMigrate(db)
+		db, _ := database.InitializePath(context.Background(), filepath.Join(dir, "bench.db"))
 		b.StartTimer()
 
 		_ = posting.UpsertAll(db, postings)
@@ -52,8 +48,7 @@ func BenchmarkPosting_UpsertAll_10k(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		b.StopTimer()
 		dir := b.TempDir()
-		db, _ := gorm.Open(sqlite.Open(filepath.Join(dir, "bench.db")), &gorm.Config{Logger: logger.Discard})
-		_ = AutoMigrate(db)
+		db, _ := database.InitializePath(context.Background(), filepath.Join(dir, "bench.db"))
 		b.StartTimer()
 
 		_ = posting.UpsertAll(db, postings)
@@ -66,8 +61,7 @@ func BenchmarkPosting_UpsertAll_50k(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		b.StopTimer()
 		dir := b.TempDir()
-		db, _ := gorm.Open(sqlite.Open(filepath.Join(dir, "bench.db")), &gorm.Config{Logger: logger.Discard})
-		_ = AutoMigrate(db)
+		db, _ := database.InitializePath(context.Background(), filepath.Join(dir, "bench.db"))
 		b.StartTimer()
 
 		_ = posting.UpsertAll(db, postings)
@@ -80,8 +74,7 @@ func BenchmarkPrice_UpsertAllByType_1k(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		b.StopTimer()
 		dir := b.TempDir()
-		db, _ := gorm.Open(sqlite.Open(filepath.Join(dir, "bench.db")), &gorm.Config{Logger: logger.Discard})
-		_ = AutoMigrate(db)
+		db, _ := database.InitializePath(context.Background(), filepath.Join(dir, "bench.db"))
 		b.StartTimer()
 
 		_ = price.UpsertAllByType(db, config.Stock, prices)